scripts/afbeelding-generator/requirements.txt
openai
supabase
python-dotenv
Pillow
